% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.GatingSetList.R
\docType{methods}
\name{clone,GatingSetList-method}
\alias{clone,GatingSetList-method}
\title{clone a GatingSetList}
\usage{
\S4method{clone}{GatingSetList}(x, ...)
}
\arguments{
\item{x}{A \code{GatingSetList}}

\item{...}{see \code{\link{clone,GatingSet-method}}}
}
\value{
A copy of a given \code{GatingSetList}.
}
\description{
A wrapper that invokes clone method for each underlying GatingSet.
}
\examples{
  \dontrun{
    gslist1<-clone(gslist)

  }
}
