% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescendants.R
\name{getDescendants}
\alias{getDescendants}
\title{Fast version of getDescendants call}
\usage{
getDescendants(gh, node)
}
\arguments{
\item{gh}{GatingHierarchy}

\item{node}{the node path}
}
\description{
Fast version of getDescendants call
}
\examples{
library(flowWorkspace)
dataDir <- system.file("extdata",package="flowWorkspaceData")
suppressMessages(gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE)))
getDescendants(gs[[1]], "CD4")
flowIncubator::getDescendants(gs[[1]], "CD4")
}

