% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\alias{getStats.GatingHierarchy}
\alias{getStats.GatingSet}
\alias{getStats.GatingSetList}
\title{Exact MFI from populations(or nodes) for all the markers}
\usage{
getStats(x, ...)

\method{getStats}{GatingSetList}(x, ...)

\method{getStats}{GatingSet}(x, ...)

\method{getStats}{GatingHierarchy}(x, nodes = NULL, type = c("count",
  "percent"), ...)
}
\arguments{
\item{x}{a GatingSet or GatingHierarchy}

\item{...}{other arguments
nodes the character vector specifies the populations of interest. default is all available nodes
type the character vector specifies the type of pop stats or
    a function used to compute population stats.
    when character, it is expected to be either "count" or "percent". Default is "count" (total number of events in the populations).
    when a function,  it takes a flowFrame object through 'fr' argument and return the stats as a named vector.
... arguments passed to \link{getNodes} method.}
}
\value{
a data.table that contains MFI values for each marker per column along with 'pop' column and 'sample' column (when used on a 'GatingSet')
}
\description{
It calculates the MFI for each marker.
}
\examples{
\dontrun{
library(flowWorkspace)
library(cytoUtils)
dataDir <- system.file("extdata",package="flowWorkspaceData")
suppressMessages(gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE)))

# get stats all nodes
dt <- getStats(gs) #default is "count"

nodes <- c("CD4", "CD8")
getStats(gs, nodes, "percent")

# pass a build-in function
getStats(gs, nodes, type = pop.MFI)

# supply user-defined stats fun
pop.quantiles <- function(fr){
   chnls <- colnames(fr)
   res <- matrixStats::colQuantiles(exprs(fr), probs = 0.75)
   names(res) <- chnls
   res
   }
getStats(gs, nodes, type = pop.quantiles)
}
}

