% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.GatingSetList.R
\name{gslist_clone}
\alias{gslist_clone}
\title{clone a GatingSetList}
\usage{
gslist_clone(x, ...)
}
\arguments{
\item{x}{A \code{GatingSetList}}

\item{...}{see \code{\link{gs_clone}}}
}
\value{
A copy of a given \code{GatingSetList}.
}
\description{
A wrapper that invokes clone method for each underlying GatingSet.
}
\examples{
  \dontrun{
    gslist1<-gslist_clone(gslist)

  }
}
