% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{readFCSPar}
\alias{readFCSPar}
\title{fast way of getting channel names from fcs file by only reading header}
\usage{
readFCSPar(fileName)
}
\arguments{
\item{fileName}{\code{character}  fcs file name(path)}
}
\value{
a \code{character} vector channels/parameters used in this FCS
}
\description{
This is a convenient wrapper around \link{read.FCSheader} and \code{flowCore:::readFCSgetPar}.
}
\examples{

readFCSPar(system.file("extdata/0877408774.B08", package = "flowCore"))
}

