% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMFI.R
\name{pop.count}
\alias{pop.MFI}
\alias{pop.count}
\title{built-in stats functions.}
\usage{
pop.count(fr)

pop.MFI(fr)
}
\arguments{
\item{fr}{a flowFrame represents a gated population}
}
\value{
a named numeric vector
}
\description{
pop.count returns the count.
pop.MFI computes and returns the median fluorescence intensity for each marker.
They are typically used as the arguments passed to \code{getStats} method to perform the sample-wise population stats calculations.
}

