% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmix2DGate.R
\name{tmix2DGate}
\alias{tmix2DGate}
\title{Gating wrapper based on flowClust::tmixFilter}
\usage{
tmix2DGate(fr, channels, K, quantile = 0.85, trans = 0, ...)
}
\arguments{
\item{fr}{\code{flowFrame}}

\item{channels}{\code{character} vector of channel/marker names (must be of length 2)}

\item{K}{\code{integer} the number of clusters to fit}

\item{quantile}{\code{numeric} vector of length K, which specifies the size of each cluster.}

\item{trans}{\code{numeric} must be 0}

\item{...}{other arguments passed to \code{tmixFilter}}
}
\value{
a \code{filters} object that contains K \code{filters} constructed from K clusters generated by flowClust
}
\description{
It basically calls flowClust to fit K clusters and construct K \code{ellipsoidGate} from them.
}
\examples{
\dontrun{
    gates <- tmix2DGate(fr = fr, channels = c("CD4", "CD8"), K = 2, quantile = 0.7)  
    autoplot(fr, "CD8","CD4") + geom_gate(gates)
}
}

