% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\alias{getStats.GatingHierarchy}
\alias{getStats.GatingSet}
\alias{getStats.GatingSetList}
\title{Exact MFI from populations(or nodes) for all the markers}
\usage{
getStats(x, ...)

\method{getStats}{GatingSetList}(x, ...)

\method{getStats}{GatingSet}(x, ...)

\method{getStats}{GatingHierarchy}(x, fun = pop.count, nodes = NULL, ...)
}
\arguments{
\item{x}{a GatingSet or GatingHierarchy}

\item{...}{other arguments
fun a function used to compute population stats. It expects 'fr' argument as a flowFrame and a named vector as return value.
    default is \code{pop.count} function.
nodes the character vector specifies the populations of interest. default is all available nodes
... arguments passed to \link{getNodes} method.}
}
\value{
a data.table that contains MFI values for each marker per column along with 'pop' column and 'sample' column (when used on a 'GatingSet')
}
\description{
It calculates the MFI for each marker.
}
\examples{
\dontrun{
library(flowWorkspace)
library(cytoUtils)
dataDir <- system.file("extdata",package="flowWorkspaceData")
suppressMessages(gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE)))

# get stats all nodes
dt <- getStats(gs) #default is "count"

# change the stats to another build-in function and specify nodes
getStats(gs, fun = pop.MFI, nodes = c("CD4", "CD8"))

# supply user-defined stats fun
pop.quantiles <- function(fr){
   chnls <- colnames(fr)
   res <- matrixStats::colQuantiles(exprs(fr), probs = 0.75)
   names(res) <- chnls
   res
   }
getStats(gs, fun = pop.quantiles, nodes = c("CD4", "CD8"))
}
}

